check for config file in incremental mode
mdadm.conf is kinda optional in incremental mode, but not really. It
would be optional if the homehost is properly encoded in the metadata of
all arrays, for then the super-minor or name could be used.

However, during the initramfs, the homehost is now known. Previously,
I worked around this by passing it via /conf/conf.d/md, but this does
not work so well now that mdadm is invoked by udev.

Therefore, mdadm needs mdadm.conf to obtain the target names (it uses
the UUID to determine that) and we cannot skip the mdadm.conf checks.

Signed-off-by: martin f. krafft <madduck@debian.org>

Conflicts:

	debian/changelog